chore(edit): renamed footer files
