+++ next-tavla/pages/[id].tsx
@@ -4,7 +4,7 @@ import { getBoardSettings } from 'utils/firebase'
 import classes from 'styles/pages/board.module.css'
 import { convertSettingsVersion } from 'utils/converters'
 import { Board } from 'Board/scenarios/Board'
-import { FooterText } from 'Board/scenarios/FooterText'
+import { FooterText } from 'Board/scenarios/Footer'
 
 export async function getServerSideProps({
     params,
