+++ next-tavla/src/Board/scenarios/Footer/index.tsx
