+++ next-tavla/src/Board/scenarios/Footer/styles.module.css
